mcs/requirements.txt
python-MCS-SDK
logging
python-dotenv